+++ lib/rubocop/cop/rspec/match_array.rb
@@ -23,9 +23,14 @@ module RuboCop
         MSG = 'Prefer `contain_exactly` when matching an array literal.'
         RESTRICT_ON_SEND = %i[match_array].freeze
 
+        # @!method match_array_with_empty_array?(node)
+        def_node_matcher :match_array_with_empty_array?, <<~PATTERN
+          (send nil? :match_array (array))
+        PATTERN
+
         def on_send(node)
-          return unless node.first_argument.array_type?
-          return if node.first_argument.percent_literal?
+          return if match_array_with_empty_array?(node)
+          return unless offensive_argument?(node)
 
           add_offense(node) do |corrector|
             array_contents = node.arguments.flat_map(&:to_a)
@@ -34,6 +39,14 @@ module RuboCop
               "contain_exactly(#{array_contents.map(&:source).join(', ')})"
             )
           end
+        end
+
+        private
+
+        def offensive_argument?(node)
+          return unless (first_argument = node.first_argument)
+
+          first_argument.array_type? && !first_argument.percent_literal?
         end
       end
     end